+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.0] - 2026-08-07
+
+### Added
+
+- `LowPassFilter` preprocessor option
+- Common `OneSidedFilter` parent class for preprocessor filters
+
+### Fixed
+
+- Properly read samples instead of "frames" from multichannel WAV files
+
 ## [0.14.0] - 2026-08-06
 
 ### Added